Create Spiderman 3 Settings Xenia Canary
Detailed settings of my working xenia manager settings for this game that enable me to be able to play with minimal problems or errors in the games lighting. This personally fixed the bright lighting issue i had during my first boot took some time but landed on these settings making it feel almost fully like the original lighting for the game and fix dark or bright light rendering.